[DRAFT] fix typo

+++ Libraries/Base/Source/Logger.cc
@@ -108,7 +108,7 @@ Color& Log::Apperance(){ return _Color; }
 
 Log& Log::operator<<(const String& value) {
   if (_Writer && _Status) {
-    auto size_of_data = value.size();
+    auto size_of_data = UInt(value.size());
     auto error = _Writer((Bytes)value.c_str(), &size_of_data);
 
     if (error != ENoError) throw Exception(error);
@@ -118,7 +118,7 @@ Log& Log::operator<<(const String& value) {
 
 Log& Log::operator<<(String&& value) {
   if (_Writer && _Status) {
-    auto size_of_data = value.size();
+    auto size_of_data = UInt(value.size());
     auto error = _Writer((Bytes)value.c_str(), &size_of_data);
 
     if (error != ENoError) throw Exception(error);
