+++ Libraries/Base/Source/Stream.cc
@@ -51,7 +51,7 @@ String Stream::Cache(){ throw Except(ENoSupport, ""); }
 
 Stream& Stream::operator<<(String&& message) {
   if (_Writer) {
-    auto size_of_data = message.size();
+    auto size_of_data = UInt(message.size());
     auto error = _Writer((Bytes)message.data(), &size_of_data);
 
     if (error != ENoError) throw Exception(error);
@@ -61,7 +61,7 @@ Stream& Stream::operator<<(String&& message) {
 
 Stream& Stream::operator<<(const String& message) {
   if (_Writer) {
-    auto size_of_data = message.size();
+    auto size_of_data = UInt(message.size());
     auto error = _Writer((Bytes)message.data(), &size_of_data);
 
     if (error != ENoError) throw Exception(error);
